+++ bindgen/src/bindings/cpp/templates/bool_conv.hpp
@@ -3,5 +3,5 @@ struct {{ typ|ffi_converter_name }} {
     static uint8_t lower({{ type_name }});
     static {{ type_name }} read(RustStream &);
     static void write(RustStream &, {{ type_name }});
-    static int32_t allocation_size({{ type_name }});
+    static uint64_t allocation_size({{ type_name }});
 };
