+++ bindgen/src/bindings/cpp/templates/bytes_conv.hpp
@@ -4,5 +4,5 @@ struct {{ ffi_converter_name|class_name }} {
     static RustBuffer lower(const {{ type_name }} &);
     static {{ type_name }} read(RustStream &);
     static void write(RustStream &, const {{ type_name }} &);
-    static int32_t allocation_size(const {{ type_name }} &);
+    static uint64_t allocation_size(const {{ type_name }} &);
 };
