+++ bindgen/src/bindings/cpp/templates/obj_conv.hpp
@@ -14,7 +14,7 @@ struct {{ typ|ffi_converter_name }} {
     static void *lower(const {{ type_name }} &);
     static {{ type_name }} read(RustStream &);
     static void write(RustStream &, const {{ type_name }} &);
-    static int32_t allocation_size(const {{ type_name }} &);
+    static uint64_t allocation_size(const {{ type_name }} &);
 private:
     {%- if obj.has_callback_interface() %}
     friend struct {{ trait_impl }};
